examples/pub_sub/scripts/requirements.txt
coloredlogs==10.0
dill==0.2.7.1
pyzmq==17.0.0
